server/django/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
PyYAML==6.0.1
psycopg2==2.9.6
django-environ==0.11.2
djoser==2.2.2.
